mtd: nand: denali: fix write_oob_data() function
The function write_oob_data() performs write access to the spare
area, so passing "false" to the 3rd argument (transfer_spare) of
denali_send_pipeline_cmd() is weird.

Actually, the transfer_spare is ignored for MAP10 operation with
SPARE_ACCESS (0x41), so this change has no impact on the hardware
behavior.  Yet, we should keep our code to make sense.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>