kbuild: rename CONFIG_SHELL to KBUILD_SHELL
All the macros with CONFIG_ prefix are supposed to be defined
in Kconfig. But CONFIG_SHELL is defined in the top Makefile,
violating the namespace rule.

Is is impossible to move CONFIG_SHELL definition to Kconfig.
Configuration rules depend on "outputmakefile" rule where
CONFIG_SHELL is already used.
(Kbuild has to know which shell to use before invoking "make config".)

So, it seems better to rename CONFIG_SHELL.
Because "SHELL" is a special variable for GNU make, we should not
use it. It looks like "KBUILD_SHELL" is a good choice.

Signed-off-by: Masahiro Yamada <yamada.m@jp.panasonic.com>
Cc: Sam Ravnborg <sam@ravnborg.org>
Cc: Michal Marek <mmarek@suse.cz>