kbuild: replace ld-version.sh with binutils-version.sh
scripts/ld-version.sh does not work for some toolchains.
For example, Linaro tool chains have a version string like this:
2.24.0.20140829

For this toolchain, scripts/ld-version.sh returns "2036482900".
This is useless.

This commit replaces it with a re-written script binutils-version.sh.
It is based on scripts/gcc-version.sh and can be used in the same way.
The script name has been changed because it can be used for not only
"ld" but also other tools such as "as", "objcopy", etc.

Signed-off-by: Masahiro Yamada <yamada.m@jp.panasonic.com>