mtd: nand: denali_dt: enable HW_ECC_FIXUP for "denali,denali-nand-dt"
There are various customizable parametes, so several variants for
this IP.  So, a generic compatible like "denali,denali-nand-dt" is
useless.  Moreover, there are multiple things wrong with this string;
"denali,denali" is redundant is one, "-dt" is unneeded as well.
(Refer to the Link tag for discussion.)

The denali_dt.c was split out and this compatible was added by Altera
for the SOCFPGA port.  So, replace it with a more specific string.

The Denali IP on SOCFPGA incorporates the hardware ECC fixup feature.
So, this capability should be associated with the compatible string.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Link: https://lkml.org/lkml/2016/12/1/450
Series-changes: 2
  - Add caps flag to of_device_id data, not hard-code in driver code
  - Delete the generic compatible.  Add Altera-specific compatible.