irqchip: gic-v3: make alloc() receive hwirq instead of hwspec
Like most of other drivers, the GIC-v3 driver calls domain_translate to
retrieve the hwirq from the fwspec.  Now, we can push the translation
code out to the framework by setting IRQ_DOMAIN_FLAG_ALLOC_HWIRQ flag.

This commit is clean-up, but what is more important is, it will allow
child irqchips in the hierarchy to pass the hwirq to GIC-v3 directly.
This means, child irqchips do not need to know (or hard-code) the
binding of GIC-v3.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>