irqchip: uniphier-aidet: receive and pass hwirq instead of fwspec
Currently, the alloc() hook of this driver receives fwspec as argument,
then decodes it to get hwirq, then encodes it into fwspec passed to
the parent.

The repeat of decode/encode of fwspec is typical in the hierarchy
irqdomain drivers.  This is useless conversion again and again, and
what is worse, child irqchips need to know the binding of the parent,
which we can know only after parsing #interrupt-cells etc.  Actually,
drivers hard-code the fwspec format of the parent (GIC), like

   parent_fwspec.param_count = 3;
   parent_fwspec.param[0] = 0;
       ...

Now, the framework allows receives and passes hwirq directly by setting
IRQ_DOMAIN_FLAG_ALLOC_HWIRQ.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>