kbuild: use .SECONDARY instead of .PRECIOUS to keep intermediate files
GNU Make automatically deletes intermediate files that are updated
in a chain of pattern rules.

Example 1) %.dts -> %.dtb -> %.dtb.S -> %.dtb.o
Example 2) %.c_shipped -> %.c -> %.o

A couple of makefiles mark such targets as .PRECIOUS to prevent Make
from deleting them, but the correct way is to use .SECONDARY.

  .SECONDARY
    Prerequisites of this special target are treated as intermediate
    files but are never automatically deleted.

  .PRECIOUS
    When make is interrupted during execution, it may delete the target
    file it is updating if the file was modified since make started.
    If you mark the file as precious, make will never delete the file
    if interrupted.

Both can avoid deletion of intermediate files, but the difference is
the behavior when Make is interrupted; .SECONDARY deletes the target,
but .PRECIOUS does not.

The use of .PRECIOUS is relatively rare since we do not want to keep
partially constructed (possibly corrupted) targets.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>