objtool: move libelf check out of top Makefile
Bernd Edlinger reports:

  The next make after an oldconfig reads in the outdated
  include/config/auto.conf which can kill the make before
  it is able to call the syncconfig target.

  $ make defconfig
  *** Default configuration is based on 'x86_64_defconfig'
  $ make
  scripts/kconfig/conf  --syncconfig Kconfig
  Makefile:936: *** "Cannot generate ORC metadata for CONFIG_UNWINDER_ORC=y,
  please install libelf-dev, libelf-devel or elfutils-libelf-devel".  Stop.

  $ sed -i s/CONFIG_UNWINDER_ORC=y// .configs
  $ make oldconfig
  Choose kernel unwinder
  > 1. ORC unwinder (UNWINDER_ORC) (NEW)
    2. Frame pointer unwinder (UNWINDER_FRAME_POINTER)
  choice[1-2?]: 2
  $ make
  Makefile:936: *** "Cannot generate ORC metadata for CONFIG_UNWINDER_ORC=y,
  please install libelf-dev, libelf-devel or elfutils-libelf-devel".  Stop.

I tried to fix it by moving the libelf evaluation to the Kconfig
stage [1], but it was rejected.

Kconfig is now able to run shell commands to evaluate any prerequisites
needed to enable CONFIG options.

For example, as shown in commit 2a61f4747eea ("stack-protector: test
compiler capability in Kconfig and drop AUTO mode"), the new Kconfig
syntax is useful to evaluate target compiler capabilities.

However, disabling the CONFIG option silently just because of missing
libelf would not be a preferred behavior; in this case, installing an
appropriate package will solve the problem. Hence, this check will be
kept in a makefile, but somewhere else than the top Makefile.

Move the check to scripts/Makefile.toolcheck so that it is run after
syncconfig.

Another behavioral change is, missing libelf for CONFIG_STACK_VALIDATION
was previously a warning, but now a error.

[1] https://patchwork.kernel.org/patch/10516049/

Reported-by: Bernd Edlinger <bernd.edlinger@hotmail.de>
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>