kconfig: set the symbol type of y, m, n to S_TRISTATE
I noticed "y", "m", and "n" are treated as S_UNKNOWN.

The .type field is not set for them, then S_UNKNOWN is implied just
because it is the first entry of 'enum symbol_type'. As a result,
expr_calc_value() compares them as strings.

This commit changes the evaluation of 'm > n'.

Previously, it was false since 'm' comes before 'n' in the alphabet.

Now, 'm > n' is true if the module symbol (CONFIG_MODULES) is enabled,
or false otherwise.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Cover-letter:
kconfig: some cleanups

END