kconfig: try best for numeric comparison in relational operator context
Kconfig files are generally written based on the assumption that
undefined symbols should be evaluated to n.

One real-world example from sound/Kconfig is:

  if !UML

Here, UML is undefined except ARCH=um. Kconfig handles most cases
nicely, in this case, '!UML' is true if UML is undefined.

Here, I want to address pitfalls for relational operators.

[Test Case 1]

  config A
          def_bool B != n

  config B
          def_bool n

This produces an empty .config

[Test Case 2]

  config A
          def_bool B != n

This produces:

  CONFIG_A=y

The difference between the two is that 'B' is undefined for the latter.
'B != n' is no longer true because the operands are compared as strings.

The difficulty is we never know the type of undefined symbols. However,
there is a big hit here. We know the right operand, 'n' is tristate.
So, it would be reasonable to assume 'B' is also tristate (or bool).
In our general assumption, undefined symbols whose type is likely
bool or tristate should be 'n'. Hence, evaluating 'B != n' as false
would be natural.

This commit changes expr_calc_value() to try its best to do numeric
comparison. At least one of the operands is non-string, compare their
numeric values.

To sum up, the behavior will be changed as follows:

  expression | before | after
  ---------------------------
    A > y    |   n    |   n
    A = y    |   n    |   n
    A < y    |   y    |   y
    A > m    |   n    |   n
    A = m    |   n    |   n
    A < m    |   y    |   y
    A > n    |   n    |   n
    A = n    |   n    |   n
    A < n    |   y    |   y
    A > 0    |   y    |   n
    A = 0    |   n    |   y
    A < 0    |   n    |   n
    A > B    |   n    |   n
    A = B    |   n    |   n
    A < B    |   y    |   y

Here, A and B are undefined symbols.

I see a gotcha in 'A < B'. Perhaps, we could assume all of undefined
symbols are bool since the use of 'string', 'int', 'hex' is rare.

At this moment, the potential problem I want to solve is 'A != n' case.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>