kbuild: stop using --include-dir MAKEFLAG
I am not a big fan of the --include-dir flag. This flag does not
become effective in the current Makefile, hence the top Makefile
needs recursion into sub-make.

Unlike VPATH, there is no usecase to search for included makefiles
in multiple locations.

Add $(srctree)/ prefix to check-in makefiles, and remove the
--include-dir flag.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>