kobject: return -ENOSPC when add_uevent_var() fails
This function never attempts to allocate memory, so returning -ENOMEM
looks weird to me. The reason of the failure is there is no more space
in the given kobj_uevent_env structure.

Let's change the error code to -ENOSPC.

This patch is safe since this function had never failed in reality.

The callers of this function put a fixed number of small strings into
the buffer.

The buffer is defined to be large enough:

  #define UEVENT_NUM_ENVP                 32      /* number of env pointers */
  #define UEVENT_BUFFER_SIZE              2048    /* buffer for the variables */

As you see WARN() in the error paths, any failure of this function is
a software bug.

If such a case had ever happened before, you would have already seen
a noisy back-trace, then you would have increased UEVENT_NUM_ENVP or
UEVENT_BUFFER_SIZE.

Nobody has ever increased UEVENT_NUM_ENVP or UEVENT_BUFFER_SIZE since
their addition, that is, this structure is always large enough.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Series-version: 2
Series-changes: 2
  - Rephrase the commit log. No code change.