kobject: return -ENOSPC when add_uevent_var() fails
This function never attempt to allocate memory, so returning -ENOMEM
looks weird to me. The reason of the failure is there is no more space
in the given kobj_uevent_env structure.

No caller of this function relies on a specific error code, so just
change it to return -ENOSPC. The intended change, if any, is the error
number displayed in log messages.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>